85609-fixing merge errors

+++ src/ESFA.DC.PeriodEnd.ReportService.Model/PeriodEnd/Common/AECApprenticeshipPriceEpisodePeriodisedValuesInfo.cs
@@ -6,7 +6,7 @@
 
         public string LearnRefNumber { get; set; }
 
-        public byte? AimSeqNumber { get; set; }
+        public int? AimSeqNumber { get; set; }
 
         public string PriceEpisodeIdentifier { get; set; }
